+++ R/check-scalar-hasElements.R
@@ -6,6 +6,7 @@
 #'
 #' @name check-scalar-hasElements
 #' @inherit params
+#' @note Updated 2019-07-29.
 #'
 #' @seealso
 #' - `assertive.properties::is_empty()`.