+++ R/check-scalar-hasInternet.R
@@ -4,6 +4,7 @@
 #'
 #' @name check-scalar-hasInternet
 #' @inherit params
+#' @note Updated 2019-07-29.
 #'
 #' @examples
 #' hasInternet()