+++ R/check-scalar-isCharacter.R
@@ -8,6 +8,7 @@
 #'
 #' @name check-scalar-isCharacter
 #' @inherit params
+#' @note Updated 2019-07-29.
 #'
 #' @seealso `assertive.strings::is_non_missing_nor_empty_character()`.
 #'