+++ R/check-vector-isMatching.R
@@ -2,6 +2,7 @@
 #'
 #' @name check-vector-isMatching
 #' @inherit params
+#' @note Updated 2019-07-29.
 #'
 #' @seealso
 #' - `grepl()`.