+++ R/package.R
@@ -4,7 +4,4 @@
 #'
 #' @keywords internal
 #'
-#' @importFrom methods as formalArgs is isGeneric selectMethod setGeneric
-#'   setMethod signature slot
-#' @importFrom utils capture.output head
 "_PACKAGE"